feat(*):Adicionando os controllers da aplicação e abstraindo o que estava nas rotas para o controller
