+++ src/modules/appointments/infra/http/controllers/AppointmentsController.ts
@@ -1,0 +1,21 @@
+import { Request, Response } from 'express';
+import { parseISO } from 'date-fns';
+import { container } from 'tsyringe';
+import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService';
+import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
+
+export default class AppointmentsController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { provider_id, date } = request.body;
+
+    const parsedDate = parseISO(date);
+
+    const createApointment = container.resolve(CreateAppointmentService);
+
+    const appointment = await createApointment.execute({
+      date: parsedDate,
+      provider_id,
+    });
+    return response.json(appointment);
+  }
+}
