+++ src/modules/users/infra/http/controllers/SessionsController.ts
@@ -1,0 +1,17 @@
+import { Request, Response } from 'express';
+import { container } from 'tsyringe';
+import AuthenticateUserService from '@modules/users/services/AuthenticateUserService';
+
+export default class SessionsController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { email, password } = request.body;
+    const authenticateUserService = container.resolve(AuthenticateUserService);
+
+    const { user, token } = await authenticateUserService.execute({
+      email,
+      password,
+    });
+    delete user.password;
+    return response.json({ user, token });
+  }
+}
