+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -2,45 +2,25 @@
 /* eslint-disable import/no-cycle */
 import { Router } from 'express';
 import multer from 'multer';
-import { container } from 'tsyringe';
 
-import CreateUserService from '@modules/users/services/CreateUserService';
-import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
 import uploadConfig from '@config/upload';
 import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
+import UsersController from '@modules/users/infra/http/controllers/UsersController';
+import UserAvatarController from '@modules/users/infra/http/controllers/UserAvatarController';
+
+const usersController = new UsersController();
+const userAvatarController = new UserAvatarController();
 
 const usersRouter = Router();
 const upload = multer(uploadConfig);
 
-usersRouter.post('/', async (request, response) => {
-  try {
-    const { name, email, password } = request.body;
-
-    const createUser = container.resolve(CreateUserService);
-
-    const user = await createUser.execute({ name, email, password });
-
-    delete user.password;
-
-    return response.json(user);
-  } catch (err) {
-    return response.status(400).json({ error: err.message });
-  }
-});
+usersRouter.post('/', usersController.create);
 
 usersRouter.patch(
   '/',
   ensureAuthenticated,
   upload.single('avatar'),
-  async (request, response) => {
-    const updateUserAvatar = container.resolve(UpdateUserAvatarService);
-    const user = await updateUserAvatar.execute({
-      user_id: request.user.id,
-      avatarFileName: request.file.filename,
-    });
-    delete user.password;
-    return response.json(user);
-  },
+  userAvatarController.update,
 );
 
 export default usersRouter;
