+++ src/shared/infra/http/routes/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import appointmentsRouter from '@modules/appointments/infra/http/routes/appointments.routes';
 import usersRouter from '@modules/users/infra/http/routes/users.routes';
-import sessionsRouter from '@shared/infra/http/routes/sessions.routes';
+import sessionsRouter from '@modules/users/infra/http/routes/sessions.routes';
 
 const routes = Router();
